server/requirements.txt
flask
flask-SQLAlchemy
flask-cors
flask_sqlalchemy
python-dotenv
requests
mysql-connector-python
gunicorn
vertexai
openai
google-cloud-texttospeech
google-cloud-translate==2.0.1